Implement flex layout to style search page
While this isn't a perfect solution, it is the best I could do.
Everything aligns with the assignment specifications, but I could not
get the input box to cooperate. This resulted in the lowest resolution
being around 470px without breaking the page. This is much higher than
the ideal 320px. I may come back to this if I figure out a potential
solution.