Create Holy Grail webpage layout with flexbox
I created the Holy Grail webpage layout as (mostly) specified in the
instructions for the assignment. The only difference I applied was using
align-content to vertically align some html instead of using flex layout
and align-items. This looks cleaner in my opinion.

My one complaint with this solution is not allowing the sidebar to
shrink. This makes the lowest-supported resolution around 550px as
opposed to the recommended 320px. However, I don't have the knowledge
needed to fix this at the moment. Perhaps I will come back and fix this
once I do.